Minor misc. improvements to UCS PR
